+++ sentry/src/main/java/io/sentry/ExternalOptions.java
@@ -43,6 +43,7 @@ public final class ExternalOptions {
   private @Nullable Boolean sendClientReports;
   private @NotNull Set<String> bundleIds = new CopyOnWriteArraySet<>();
   private @Nullable Boolean enabled;
+  private @Nullable Boolean enablePrettySerializationOutput;
 
   @SuppressWarnings("unchecked")
   public static @NotNull ExternalOptions from(
@@ -117,6 +118,9 @@ public final class ExternalOptions {
     options.setIdleTimeout(propertiesProvider.getLongProperty("idle-timeout"));
 
     options.setEnabled(propertiesProvider.getBooleanProperty("enabled"));
+
+    options.setEnablePrettySerializationOutput(
+        propertiesProvider.getBooleanProperty("enable-pretty-serialization-output"));
 
     for (final String ignoredExceptionType :
         propertiesProvider.getList("ignored-exceptions-for-type")) {
@@ -357,5 +361,14 @@ public final class ExternalOptions {
 
   public void setEnabled(final @Nullable Boolean enabled) {
     this.enabled = enabled;
+  }
+
+  public @Nullable Boolean isEnablePrettySerializationOutput() {
+    return enablePrettySerializationOutput;
+  }
+
+  public void setEnablePrettySerializationOutput(
+      final @Nullable Boolean enablePrettySerializationOutput) {
+    this.enablePrettySerializationOutput = enablePrettySerializationOutput;
   }
 }